--- a/package.json
+++ b/package.json
@@ -25,7 +25,6 @@
     "fruitdown": "1.0.1",
     "inherits": "2.0.1",
     "js-extend": "1.0.1",
-    "level-sublevel": "6.5.4",
     "level-write-stream": "1.0.0",
     "levelup": "1.3.1",
     "lie": "3.0.2",
@@ -36,6 +35,7 @@
     "request": "2.69.0",
     "scope-eval": "0.0.3",
     "spark-md5": "2.0.0",
+    "sublevel-pouchdb": "1.0.0",
     "through2": "2.0.0",
     "vuvuzela": "1.0.2"
   },