I forked [level-sublevel](http://github.com/dominictarr/level-sublevel) into [sublevel-pouchdb](http://github.com/nolanlawson/sublevel-pouchdb) and removed what we weren't using, including a lot of extra dependencies from `levelup` itself that were adding to the Browserify size. As a result, our level-based plugins have gone from 192K minified to 184K (gzipped: 56.2K -> 53.7K).

This also increases our performance a bit, because I was able to optimize two of the hot functions (`decode`/`encode` in `sublevel/codec/legacy.js`) which had some unnecessary checks for non-string keys and complex keys. Since all of our keys are just a single String prefix, I could replace all the array operations with basic string concatenation.

As a result, a `PERF=1` run using `LEVEL_ADAPTER=memdown` on Node 5.3.0 went from 24.67s to 22.65s. The biggest gain comes from reading the database as a stream (i.e. `allDocs()`); I saw that the `all-docs-skip-limit` test went from 4.726s to 3.763s. Upping the iterations on that test from 50 to 500 for better granularity, I got 41.080s -> 29.571s (28% faster), which is pretty awesome.

Keep in mind that all the above perf numbers are for the MemDOWN adapter, so the difference will probably be less stark for LevelDB/localstorage-down/fruitdown.