@nolanlawson 2016-01-31
Here's where you can see that `legacy.js` was taking up ~2.5% of self-time for our perf tests (using MemDOWN):

![screenshot 2016-01-31 14 28 55](https://cloud.githubusercontent.com/assets/283842/12704337/04fb245a-c827-11e5-86f7-efe7df009e2d.png)

[cpuprofile file](https://nolanlawson.s3.amazonaws.com/www/cpuprofile/memdown-perf-basics-31012016.cpuprofile)

And here's the change I made:

![screenshot 2016-01-31 14 31 03](https://cloud.githubusercontent.com/assets/283842/12704345/486e964a-c827-11e5-9331-25483e09ff15.png)