@nolanlawson 2016-02-01
Well, the reason I didn't contribute back to level-sublevel was a few things:

1. The format we currently use is known as the "legacy" format. It will be unsupported at some point, and we have no burning need to upgrade because it's fine for our use case (we don't need anything as complex as [bytewise](https://www.npmjs.com/package/bytewise) just to prefix a string.
2. My "fixes" were mostly about removing code PouchDB wasn't using, so not much of value to `level-sublevel` there.
3. My fix to remove levelup is not really consistent with what level-sublevel themselves will probably want to do, which is upgrade to the `level-codec` and `level-iterator-stream` modules (which were recently pulled out levelup). In this case, I just pulled the code out of levelup itself and put it in sublevel-pouchdb, because it was easier that way.

Again, `sublevel-pouchdb` is mostly about solving the problem of having a layer over levelup with prefxied keys (in our legacy format). That's it. It's very particular to PouchDB's use case and probably not of much interest to anyone else.