@nickcolley 2016-02-01
That's totally reasonable, sorry my thoughts were forgetting the legacy nature of the package.

Hope that's okay :stuck_out_tongue: 

I'm :+1: to this otherwise!