@daleharvey 2016-02-03
So this is a fairly complex dependency that we have given ourselves the maintainance burden of, by splitting it out into its own package at least we wont get bitten by upgrades but I do think long term we should be looking to migrate to using the plain leveldown api as I think sublevel is bringing quite a bit of complexity that we dont need.

:+1: to this, I think we should move `sublevel-pouchdb` to the pouchdb org before merging though.