@nolanlawson 2016-02-05
I agree, it's a bit of a risk to take the maintenance burden on ourselves, but as-is this is just a very thin layer over vanilla levelup. And the possibility to remove the dependency entirely is still there, although TBH I'm not sure about the value of that, because this actually provides some nice abstractions that make it easier e.g. to write the `LevelTransaction` object.

I'll move `sublevel-pouchdb` to the PouchDB org and then merge this.